Add(.user): web adapter
